# server/docs/swagger.mergedv3.yaml
openapi: 3.0.3
info:
  title: Универсальный API каталога интернет-магазина
  version: 2.4.0
  description: >
    Канонический API для универсального шаблона интернет-магазина (modular
    monolith friendly).


    Концепции (DSL e-commerce):  - Category: дерево категорий неограниченной
    глубины (parentId/path/ancestors)  - ProductGroup: одна SEO-страница товара
    (карточка товара)
      - Offer: SKU/вариация товара, описана optionValues[] строго в порядке ProductGroup.variationAxes[]
      - Characteristics: динамические характеристики товара (ProductGroup.characteristics[])
      - Characteristics Meta: словарь характеристик для построения фильтров на фронте (type/unit/valuesPreset/filterable/searchable)
      - Catalog Cards: быстрый DTO для листинга (агрегация цен, наличия, preview по вариациям)
      - Soft-deletes/visibility: status=active/hidden/draft (без удаления из БД)
      - Multi-category: ProductGroup может принадлежать нескольким категориям (categoryIds[])
servers:
  - url: /v1
...
components:
  securitySchemes:
    PublicApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        Опционально. Используется если нужно ограничить публичный доступ или
        сделать rate-limit по ключам.

        Если у проекта публичный каталог — схему можно не включать на уровне
        операций.
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    LocalizedText:
      type: object
      additionalProperties: true
      example:
        ua: Назва товару
        en: Название товара
...
    UserShort:
      type: object
      properties:
        id:
          type: string
          example: 67f2c5d1b9e12a0012345678
        email:
          type: string
          format: email
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        middleName:
          type: string
          nullable: true
        birthDate:
          type: string
          format: date-time
          nullable: true
        phone:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - active
            - blocked
            - pending
        bonusBalance:
          type: number
        referralBalance:
          type: number
        referralCode:
          type: string
          nullable: true
        invitedBy:
          type: string
          nullable: true
        deliveryProvince:
          type: string
          nullable: true
        deliveryCity:
          type: string
          nullable: true
        deliveryPostOffice:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserShort'
        token:
          type: string
        expiresIn:
          type: integer
      required:
        - user
        - token
        - expiresIn
    AuthMeResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        middleName:
          type: string
          nullable: true
        birthDate:
          type: string
          format: date-time
          nullable: true
        phone:
          type: string
          nullable: true
        deliveryProvince:
          type: string
          nullable: true
        deliveryCity:
          type: string
          nullable: true
        deliveryPostOffice:
          type: string
          nullable: true
        status:
          type: string
        bonusBalance:
          type: number
        referralBalance:
          type: number
        referralCode:
          type: string
          nullable: true
        invitedBy:
          type: string
          nullable: true
        cart:
          type: array
          items:
            $ref: '#/components/schemas/CartItemRaw'
        wishlist:
          type: array
          items:
            $ref: '#/components/schemas/ProductGroup'
        orders:
          type: array
          items:
            $ref: '#/components/schemas/UserOrderRef'
        createdAt:
          type: string
          format: date-time
    UserOrderRef:
      type: object
      properties:
        orderId:
          type: string
        orderNumber:
          type: number
        createdAt:
          type: string
          format: date-time
        total:
          type: number
        status:
          type: string
    ProductGroup:
      type: object
      additionalProperties: true
      description: |
        ProductGroup отдается как есть из Mongo. Структура может быть шире,
        чем описано здесь.
      properties:
        _id:
          type: string
        title:
          $ref: '#/components/schemas/LocalizedText'
        name:
          $ref: '#/components/schemas/LocalizedText'
        slug:
          type: string
        reviews:
          type: array
          items:
            type: object
    CartItemRaw:
      type: object
      properties:
        offerId:
          type: string
        qty:
          type: integer
        priceAtAdd:
          type: number
        groupId:
          type: string
          nullable: true
        sku:
          type: string
          nullable: true
...
    OrderItem:
      type: object
      properties:
        offerId:
          type: string
        groupId:
          type: string
          nullable: true
        sku:
          type: string
          nullable: true
        optionKey:
          type: string
          nullable: true
        optionValues:
          type: array
          items: {}
        titleSnapshot:
          nullable: true
        imgSnapshot:
          type: string
          nullable: true
        qty:
          type: integer
        pricePerUnit:
          type: number
        subtotal:
          type: number
    Payment:
      type: object
      properties:
        payment_method_id:
          type: integer
        payment_method:
          type: string
        amount:
          type: number
        description:
          type: string
          nullable: true
        status:
          type: string
        invoiceId:
          type: string
          nullable: true
        pageUrl:
          type: string
          nullable: true
        monoPartsOrderId:
          type: string
          nullable: true
        partsCount:
          type: integer
          nullable: true
        sitniksOrderId:
          type: integer
          nullable: true
        monoPartsConfirmed:
          type: boolean
        monoPartsConfirmedAt:
          type: string
          format: date-time
          nullable: true
        monoPartsConfirmResponse:
          nullable: true
    SitniksInfo:
      type: object
      properties:
        orderId:
          type: integer
          nullable: true
        statusId:
          type: integer
          nullable: true
        statusCode:
          type: string
          nullable: true
        statusTitle:
          type: string
          nullable: true
        lastSyncAt:
          type: string
          format: date-time
          nullable: true
    Order:
      type: object
      properties:
        _id:
          type: string
        orderNumber:
          type: integer
        order_number:
          type: string
        payments:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        userId:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        subtotal:
          type: number
        usedBonusBalance:
          type: number
        usedReferralBalance:
          type: number
        totalToPay:
          type: number
        finalPrice:
          type: number
          nullable: true
        currency:
          type: string
          example: UAH
        lastName:
          type: string
        firstName:
          type: string
        middleName:
          type: string
          nullable: true
        customerPhone:
          type: string
        customerEmail:
          type: string
          format: email
        deliveryProvince:
          type: string
        deliveryCity:
          type: string
        deliveryPostOffice:
          type: string
        promoCode:
          type: string
        status:
          type: string
          enum:
            - pending_payment
            - paid
            - processing
            - shipped
            - delivered
            - cancelled
            - installment_waiting_store_confirm
        payment:
          type: string
        installmentMonths:
          type: integer
        sitniks:
          $ref: '#/components/schemas/SitniksInfo'
        createdAt:
          type: string
          format: date-time
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - firstName
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        firstName:
          type: string
        ref:
          type: string
          nullable: true
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required:
        - token
        - password
      properties:
        token:
          type: string
        password:
          type: string
          format: password
    UpdateAuthMeRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
          nullable: true
        birthDate:
          type: string
          format: date
          nullable: true
        phone:
          type: string
        deliveryProvince:
          type: string
        deliveryCity:
          type: string
        deliveryPostOffice:
          type: string
...
paths:
...
  /auth/register:
    post:
      tags:
        - Auth
      summary: Регистрация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Пользователь зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: Пользователь уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Auth
      summary: Логин пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Нет email или password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Пользователь не активен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags:
        - Auth
      summary: Получить текущего пользователя
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Данные текущего пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMeResponse'
        '401':
          description: Нет токена или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - Auth
      summary: Обновить профиль текущего пользователя
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAuthMeRequest'
      responses:
        '200':
          description: Пользователь обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserShort'
        '401':
          description: Нет токена или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/forgot-password:
    post:
      tags:
        - Auth
      summary: Запрос на сброс пароля
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Ответ без раскрытия существования email
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Email обязателен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/reset-password:
    post:
      tags:
        - Auth
      summary: Сброс пароля по токену
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Пароль обновлен
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Невалидный запрос или токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /user/orders:
    get:
      tags:
        - Orders
      summary: Получить список заказов пользователя
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список заказов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...